🚚 rename: timeUtils 이름을 calculateLeftTime로 변경 및 타입 분리 #93
